♻️  Move fixture to __init__ and add new model test
